gym_hunter/requirements.txt
google-cloud-aiplatform>=1.38.0
googlemaps>=4.10.0
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
